worker: monitor shuts down worker prematurely
fix: #77
This bug is because in our heartbeat detecting the `inspect.schedule` and
`inspect.reserved` didn't return current tasks in queue.  The
`inspect.reserved` it will lists tasks that had been prefetched, but we
had disabled this prefetch behavior. It is always empty for now. And the
prematurely shutdown is caused when the it meets the time gap between
two tasks. Here I directly go into the queue directory and return all
of the tasks in the queue to solve this problem.